# openapi.yaml
openapi: 3.1.0
info:
  title: RMM API
  description: |
    Remote Management System API for managing clients in a local network.
    All endpoints require authentication via the `X-API-Key` header.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    HeartbeatRequest:
      type: object
      required:
        - client_id
        - hostname
        - os
        - arch
        - uptime_seconds
      properties:
        client_id:
          type: string
          description: Unique client identifier (defaults to hostname)
          example: workstation-01
        hostname:
          type: string
          description: Machine hostname
          example: workstation-01
        os:
          type: string
          description: Operating system (from Go runtime.GOOS)
          example: linux
        arch:
          type: string
          description: CPU architecture (from Go runtime.GOARCH)
          example: amd64
        uptime_seconds:
          type: number
          format: double
          description: Seconds since the client process started
          example: 3600.5

    HeartbeatResponse:
      type: object
      properties:
        tasks:
          type: array
          description: Pending tasks for the client to execute (may be empty or absent)
          items:
            $ref: '#/components/schemas/Task'

    Task:
      type: object
      required:
        - type
        - payload
      properties:
        id:
          type: string
          description: Unique task identifier. Auto-generated by the server if omitted.
          example: task-1718000000000000000
        type:
          type: string
          description: Task type that determines how the payload is interpreted
          enum:
            - exec
            - resource
          example: exec
        payload:
          type: object
          description: |
            Key-value pairs specific to the task type.

            **exec**: Run a shell command.
            - `command` (required): The shell command to execute.

            **resource**: Query a system resource.
            - `name` (required): One of `hostname`, `os`, `pid`, `env`.
            - `key` (required for `name=env`): The environment variable name.
          additionalProperties:
            type: string
          example:
            command: "uptime"

    TaskResult:
      type: object
      required:
        - task_id
        - client_id
        - output
        - exit_code
      properties:
        task_id:
          type: string
          description: ID of the task this result belongs to
          example: task-1718000000000000000
        client_id:
          type: string
          description: ID of the client that executed the task
          example: workstation-01
        output:
          type: string
          description: Combined stdout/stderr output of the executed task
          example: " 10:00:01 up 1 day,  2:30,  1 user,  load average: 0.10, 0.08, 0.05\n"
        exit_code:
          type: integer
          description: Exit code of the process. 0 indicates success, any other value indicates failure.
          example: 0
        error:
          type: string
          description: Error message if the task could not be executed at all (e.g. binary not found). Omitted on success.
          example: "execution failed: exit status 1"

    ClientRecord:
      type: object
      properties:
        client_id:
          type: string
          example: workstation-01
        hostname:
          type: string
          example: workstation-01
        os:
          type: string
          example: linux
        arch:
          type: string
          example: amd64
        uptime_seconds:
          type: number
          format: double
          example: 3600.5
        last_contact:
          type: string
          format: date-time
          description: Timestamp of the last received heartbeat (UTC)
          example: "2024-06-10T10:00:01Z"
        online:
          type: boolean
          description: True if last contact was within the past 2 minutes
          example: true

    Error:
      type: object
      properties:
        message:
          type: string
          example: Client not found

paths:
  /heartbeat:
    post:
      summary: Client check-in
      description: |
        Called periodically by each client to report its status.
        The server upserts the client record and returns any pending tasks.
      operationId: heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
            example:
              client_id: workstation-01
              hostname: workstation-01
              os: linux
              arch: amd64
              uptime_seconds: 3600.5
      responses:
        '200':
          description: Heartbeat accepted. Response contains pending tasks (may be empty).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
              examples:
                no_tasks:
                  summary: No pending tasks
                  value:
                    tasks: []
                with_tasks:
                  summary: One pending task
                  value:
                    tasks:
                      - id: task-1718000000000000000
                        type: exec
                        payload:
                          command: "uptime"
        '400':
          description: Invalid request body or missing client_id
        '401':
          description: Missing or invalid API key

  /task-result:
    post:
      summary: Submit task result
      description: |
        Called by the client after executing a task received in a heartbeat response.
        The server logs and persists the result, which can then be retrieved via
        `GET /tasks/{id}/result`.
      operationId: submitTaskResult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskResult'
            examples:
              success:
                summary: Successful task result
                value:
                  task_id: task-1718000000000000000
                  client_id: workstation-01
                  output: " 10:00:01 up 1 day,  2:30,  1 user\n"
                  exit_code: 0
              failure:
                summary: Failed task result
                value:
                  task_id: task-1718000000000000001
                  client_id: workstation-01
                  output: "sh: unknown_cmd: command not found\n"
                  exit_code: 127
      responses:
        '204':
          description: Result received successfully
        '400':
          description: Invalid request body
        '401':
          description: Missing or invalid API key

  /clients:
    get:
      summary: List all clients
      description: |
        Returns all known clients. The `online` field is computed dynamically:
        a client is considered online if its last heartbeat was within the past 2 minutes.
      operationId: listClients
      responses:
        '200':
          description: List of all registered clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientRecord'
              example:
                - client_id: workstation-01
                  hostname: workstation-01
                  os: linux
                  arch: amd64
                  uptime_seconds: 3600.5
                  last_contact: "2024-06-10T10:00:01Z"
                  online: true
                - client_id: laptop-02
                  hostname: laptop-02
                  os: windows
                  arch: amd64
                  uptime_seconds: 120.0
                  last_contact: "2024-06-10T08:00:00Z"
                  online: false
        '401':
          description: Missing or invalid API key

  /tasks/{id}/result:
    get:
      summary: Retrieve a task result
      description: |
        Returns the result of a completed task by its ID.
        Results are held in memory and available as soon as the client submits them
        via `POST /task-result`. Returns 404 if the task has not yet completed
        or the ID is unknown.
      operationId: getTaskResult
      parameters:
        - name: id
          in: path
          required: true
          description: The task ID to retrieve the result for
          schema:
            type: string
          example: task-1718000000000000000
      responses:
        '200':
          description: Task result found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResult'
              examples:
                success:
                  summary: Successful task result
                  value:
                    task_id: task-1718000000000000000
                    client_id: workstation-01
                    output: " 10:00:01 up 1 day,  2:30,  1 user\n"
                    exit_code: 0
                failure:
                  summary: Failed task result
                  value:
                    task_id: task-1718000000000000001
                    client_id: workstation-01
                    output: "sh: unknown_cmd: command not found\n"
                    exit_code: 127
                    error: ""
        '401':
          description: Missing or invalid API key
        '404':
          description: No result found for this task ID (task unknown or not yet completed)

  /clients/{id}/task:
    post:
      summary: Enqueue a task for a client
      description: |
        Adds a task to the client's queue. The task will be delivered to the client
        on its next heartbeat. Only one batch of tasks is held per client at a time.

        **Task types:**
        - `exec` – runs a shell command (`sh -c` on Linux/macOS, `cmd /C` on Windows)
        - `resource` – queries a system value without running a shell
      operationId: enqueueTask
      parameters:
        - name: id
          in: path
          required: true
          description: The `client_id` of the target client
          schema:
            type: string
          example: workstation-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
            examples:
              exec_task:
                summary: Run a shell command
                value:
                  type: exec
                  payload:
                    command: "df -h"
              resource_hostname:
                summary: Query hostname resource
                value:
                  type: resource
                  payload:
                    name: hostname
              resource_env:
                summary: Query an environment variable
                value:
                  type: resource
                  payload:
                    name: env
                    key: PATH
      responses:
        '202':
          description: Task accepted and enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    description: The ID of the enqueued task, use this to retrieve the result later
                    example: task-1718000000000000000
        '400':
          description: Invalid request body
        '401':
          description: Missing or invalid API key
        '404':
          description: Client not found